% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.EEZarea.r
\name{get.EEZarea}
\alias{get.EEZarea}
\title{This is a function to assign the EEZ area (country or territory) given longitude and the latitude
It uses the shapefile from MarineRegions.org (Flanders Marine Institute)}
\usage{
get.EEZarea(londec, latdec, type = c("country"))
}
\arguments{
\item{londec}{Longitude in decimal degrees}

\item{latdec}{Latitude in decimal degrees}
}
\value{
Returns a vector of EEZs.
}
\description{
This is a function to assign the EEZ area (country or territory) given longitude and the latitude
It uses the shapefile from MarineRegions.org (Flanders Marine Institute)
}
\examples{
get.EEZ(londec=c(-18,12.588,13.218,16,-15,5,0,-42,-21,-18.3,-24.6,-18),
latdec=c(22,-11.2237,-22.99078,-85,-22,-40,-13,15,-15,45,-30,28),type='country')
get.EEZ(londec=c(-18,12.588,13.218,16,-15,5,0,-42,-21,-18.3,-24.6,-18),
latdec=c(22,-11.2237,-22.99078,-85,-22,-40,-13,15,-15,45,-30,28),type='territory')
}
