% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.FAOarea.R
\name{get.FAOarea}
\alias{get.FAOarea}
\title{Esta funcion asigna una zona FAO dadas las coordenadas.}
\usage{
get.FAOarea(lon, lat, f_level = "MAJOR")
}
\arguments{
\item{lon}{vector of longitudes}

\item{lat}{vector of latitudes}

\item{f_level}{c('MAJOR','SUBAREA','DIVISION','SUBDIVISION','SUBUNIT') Type of FAO area, being "MAJOR" the default.}
}
\description{
Esta funcion asigna una zona FAO dadas las coordenadas.
}
\examples{
get.FAOarea(lon=c(-50,-30,-20,0),lat=c(20,-20,10,-20),f_level='MAJOR')
}
