% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.ICCATarea.R
\name{get.ICCATarea}
\alias{get.ICCATarea}
\title{This is a function to assign the ICCAT statistical area and the stock.
given the 3-letter FAO code, the longitude and the latitude}
\usage{
get.ICCATarea(faocode, londec, latdec)
}
\arguments{
\item{faocode}{Species FAO code}

\item{londec}{Longitude in decimal degrees}

\item{latdec}{Latitude in decimal degrees}

\item{plot}{T or F}
}
\value{
Returns a matrix with two columns (statistical areas and stocks).
}
\description{
This is a function to assign the ICCAT statistical area and the stock.
given the 3-letter FAO code, the longitude and the latitude
}
\examples{
get.ICCATarea(faocode=c('SKJ','SKJ','SKJ','SKJ','YFT','YFT','YFT','TOT','TOT','SWO','SWO'),
londec=c(-20,-22.5,-22.5,160,-15,5,0,-42,-21,-18.3,-24.6),
latdec=c(-15,25,25,-85,-22,-40,-13,15,-15,45,-30))
}
